RFC: Don't redirect when loading files index page
Currently we are redirecting from ?dir=/&fileid=2 to ?dir=/. This is an
issue because we then need to load two pages with full file system setup
and authentification instead of one and the assets won't start loading
until the second page is delivered to the user.

Additionally when loading ?dir=/, we then change the url back to
?dir=/&fileid=2 (without reload) so that the next time we load the page
again we do the same thing again.

Depending on the speed of the server and internet connection we can save
100ms to 400ms, improving the user experience.

Signed-off-by: Carl Schwan <carl@carlschwan.eu>